requirements.txt
Mako
MarkupSafe
PyMySQL
atomicwrites
attrs
certifi
chardet
colorama
configparser
crayons
glob2
idna
jsonpickle
more-itertools
mysql-connector-python-rf
parse
parse-type
pip
pluggy
pony
py
pytest
pytest-bdd
requests
selenium
setuptools
six
urllib3
webdriver-manager
wheel
lxml-4.3.3-cp37-cp37m-win32.whl
allure-pytest